+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - PostgreSQL, MS SQL: Show two columns in select COUNT(*), COUNT(DISTINCT id)
 - MS SQL: Treat binary fields as blob (bug SF-369)
 - Elasticsearch, ClickHouse: Use default port
+- MongoDB: Authenticate against the database used in login
 - SimpleDB: Use default server
 - Editor: Display all images recognized by PHP, set their size, load them lazily
 - Plugins: Method verifyVersion()